Compile-time argument KERNEL_TYPE
Previously the data type for the kernel was split into two parts in order to specify vector sizes. But this hinders the ability to use a vector type for the input image, hence the reason why this is simplified into a single argument. That way the filter kernel can be either a scalar or vector type.